doubly fine-tuned models & other sweep improvements & config files (#112)
* added compliance overview plot

* setting default non-determinacy to 5

* added hacky script to trawl the logs and pull experimental cost

* fix inference override dict parsing

* almost all configs from new run

* final model config for first pass on jun20 sweep

* Claude 3.5 Sonnet config

* commented out logging code that makes a matplotlib image for every call (thanks James!)

* updated hardcoded models in Anthropic API

* added pricing for Claude 3.5 Sonnet

* treating Gemini returning 0 candidates as safety refusal => empty string

* jun20 second pass model cfgs

* model divergent strings extraction doesn't overwrite by default

* models for half held out

* restarting processes that die with error code 137 or 139

* added helper function for canonical safe model name transforms

* more doubly finetuned configs

* retry with truncated messages if the OpenAI API complains due to Gemini repetition (WIP)

* fix too long study name for wandb

* added `doubly_trained_model_configs` argument to train models again

---------

+++ evals/extract_model_divergent_strings.py
@@ -21,7 +21,7 @@ from evals.analysis.loading_data import get_data_path, load_and_prep_dfs  # noqa
 
 
 def extract_most_uncertain_strings_from_base(
-    input_file_paths, n_out_strings=float("inf"), output_file_path=None, response_properties=["identity"]
+    input_file_paths, n_out_strings=float("inf"), output_file_path=None, response_properties=["identity"], overwrite=False
 ) -> Path:
     """Extracts the strings that different models generate different base completions for. Saves a .csv with the strings into the same directory as the input file called `out_strings.csv`.
 
@@ -30,6 +30,7 @@ def extract_most_uncertain_strings_from_base(
         n_out_strings: Number of strings to extract. If we can't find enough strings, we will return all the strings we have.
         output_file_path: Path to save the output file. If None, will save to the same directory as the input file.
         response_properties: List of response properties to compare. Default is ['identity']. The strings that are different for **all** of the response properties will be included.
+        overwrite: If True, will overwrite the output file if it already exists.
 
     Returns:
         Path to the .csv with the extracted strings.
@@ -89,6 +90,9 @@ def extract_most_uncertain_strings_from_base(
     # ensure output dir exists
     output_file_path = Path(output_file_path)
     output_file_path.parent.mkdir(parents=True, exist_ok=True)
+    if output_file_path.exists() and not overwrite:
+        LOGGER.warning(f"Model divergent strings at {output_file_path} already exist. Not overwriting.")
+        return output_file_path
 
     out_df[["string"]].to_csv(output_file_path, index=False, quoting=csv.QUOTE_ALL)
     LOGGER.info(f"Saved {len(out_df)} strings to {output_file_path}")
